Add config.application_url configuration option for rails.
You can set a application_url in your config.yml file.
It is available as a ActionDispatch::URI class under Rails.application.url

It uses Rubys Core URI class and adds some rails specific methods to it.
Some parts of the URL class moved their and hopefully we can add more.

We then use the Rails.application.url and preconfigure action mailer and
controllers.

**Some remarks for further refactoring work
Url and uri are very close. Url is build around a request. URI is a wrapper for the ruby URI Api.
This is a first step to draw the line sharper, more work should follow.
In the long term Url should be best called RequestUrl and concentrate on
rails and request specific use cases. And URI should just be a minimal
wrapper around Ruby URI

One of the quirks that make refactoring and simplifying this
harder is, at least in test cases, the url objects gets initalized -
but the path is added in a later step.