cleanup action_dispatch/http folder
Following common rules this declutters the http folder with all the different modules/classes mixed together

- Move ActionDispatch::Response out of action_dispatch/http and the included FilterRedirect into ActionDispatch::Response.
Move modules that are part of request from action_dispatch/http to action_dispatch/request folder.

- Separate internal classes in ActionDIspatch::Response
as ActionDIspatch::Response is complex and it is hard to get an overview - I moved 3 classes inside of it to its own file.